requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
langgraph==0.0.26
langchain==0.1.4
langchain-groq==0.0.1
supabase==2.3.0
sentence-transformers==2.3.1
httpx==0.26.0
pytest==7.4.4
